Packaging documentation
Please test, will do the same on my side.
